refactor: 코어 함수 typescript로 이미그레이션
